Remove unrelocatable bits
These can be relocated on the ROM because they are at addresses such as
0x000000ab, but in the flash they end up at addresses such as 0x100000ab
which is too big to relocate.

We don't need the external ROM function table (the __memcpy and the
like) anyway. But maybe we could use the original USB rom functions
table (the _msc_cmd_packet and the like) instead?

Also ensuring that the USB IRQ __irq5_vector ends up at the correct
place by using `.org` -- maybe we could point it at the original
isr_irq5 instead though?