Don't abort early on flash boot
Instead we reset core to actual ROM USB boot on an attempt to write to
flash.

This is so that if someone tries to write to the flash, we don't
overwrite ourselves, but still give a way for them to program the
RP2040.

It would be better to do actual hand-over but this is fine for now.